Ensure categories are saved in the model when updated
Without this, moving a launcher into a custom directory results in
the launcher disappearing from the menu structure for the usual menu
program/when menulibre restarts.

Launchers in a custom directory require an explicit include to
appear, but when the associated category hasn't been saved in the
model, model_to_xml_includes doesn't realise the launcher is in a
custom dir so it only ends up in the relevant Layout section,
therefore disappearing.

Fixes https://bugs.launchpad.net/menulibre/+bug/1746802